adds parallelism to Directory and Workspace (#529)
* adds parallelism to Directory and Workspace

Adds a parallelism field to Directory and Workspace entities. This controls the number of
concurrent operations performed by a terraform apply in a Workspace. When a
Workspace is created, it's parallelism will be set to that of it's parent Directory, if one is
set. Otherwise, the default Terraform value (currently 10) is used.